internals/scaffold/templates/fastapi/requirements.txt
fastapi[standard]===0.116.1
python-dotenv===1.1.1
